# gateway/openapi.yaml
openapi: "3.1.0"
info:
  title: Sentinel Gateway API
  version: "1.0.0"
  description: |
    Real-time AI agent governance layer. Intercepts tool calls, evaluates them
    against policies, and returns allow/block/review decisions with a full audit trail.
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://sentinel.internal
    description: Production

# =============================================================================
# Security
# =============================================================================
security:
  - AgentApiKey: []
components:
  securitySchemes:
    AgentApiKey:
      type: apiKey
      in: header
      name: X-Sentinel-Agent-Key
      description: >
        API key issued at agent registration. Format: `snl_<hex>`.
        Exchange for a short-lived JWT via POST /v1/auth/token.
    BearerJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ---------------------------------------------------------------------------
  # Shared schemas
  # ---------------------------------------------------------------------------
  schemas:
...
    Verdict:
      type: string
      enum: [ALLOWED, BLOCKED, HUMAN_REVIEW]
    DecisionPath:
      type: string
      enum: [fast_path, cognitive_path]

    # --- Core Request/Response -----------------------------------------------
...
    DecisionResponse:
      type: object
      required: [decision_id, verdict, path, latency_ms, policy_version]
      properties:
        decision_id:
          type: string
          example: dec_01hx2k3m4n5p
        verdict:
          $ref: "#/components/schemas/Verdict"
        reason:
          type: string
          description: Human-readable explanation of the decision.
          example: "Blocked: recipient matches denylist pattern admin@"
        path:
          $ref: "#/components/schemas/DecisionPath"
        latency_ms:
          type: number
          format: float
          example: 7.3
        policy_version:
          type: string
          description: Semver of the active policy that evaluated this request.
          example: "data-access-v3.1.0"
        confidence:
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
          description: Judge confidence (cognitive path only).
        rate_limit:
          $ref: "#/components/schemas/RateLimitInfo"
    PendingDecisionResponse:
      type: object
      description: Returned with HTTP 202 when cognitive path exceeds sync budget.
      required: [decision_id, status, poll_url, estimated_ms]
      properties:
        decision_id:
          type: string
          example: dec_01hx2k3m4n5p
        status:
          type: string
          enum: [pending]
        poll_url:
          type: string
          example: /v1/decisions/dec_01hx2k3m4n5p
        estimated_ms:
          type: integer
          example: 5000
    RateLimitInfo:
      type: object
      properties:
        tokens_remaining:
          type: integer
          example: 45
        tokens_max:
          type: integer
          example: 50
        reset_at:
          type: string
          format: date-time

    # --- Agent Registration --------------------------------------------------
    AgentRegistrationRequest:
      type: object
      required: [name, policy_group]
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 64
          example: financial_analyst_agent
        policy_group:
          type: string
          description: Which policy group governs this agent's tool calls.
          example: financial
        tenant_id:
          type: string
          description: Optional. Defaults to "default" tenant.
          example: acme_corp
        metadata:
          type: object
          description: Arbitrary agent metadata (version, owner, etc.)
    AgentRegistrationResponse:
      type: object
      required: [agent_id, api_key, name, policy_group, registered_at]
      properties:
        agent_id:
          type: string
          example: agent_01hx2k3m
        api_key:
          type: string
          description: >
            Store this securely — it will NOT be shown again.
          example: snl_a3f8b2c1d4e5f6...
        name:
          type: string
        policy_group:
          type: string
        tenant_id:
          type: string
        registered_at:
          type: string
          format: date-time
    AgentInfo:
      type: object
      properties:
        agent_id:
          type: string
        name:
          type: string
        policy_group:
          type: string
        tenant_id:
          type: string
        registered_at:
          type: string
          format: date-time
        total_requests:
          type: integer
        last_active:
          type: string
          format: date-time

    # --- Auth ----------------------------------------------------------------
    TokenRequest:
      type: object
      description: Exchange an API key for a short-lived JWT.
      required: [agent_id]
      properties:
        agent_id:
          type: string
          example: agent_01hx2k3m
    TokenResponse:
      type: object
      required: [access_token, expires_in, token_type]
      properties:
        access_token:
          type: string
        token_type:
          type: string
          enum: [Bearer]
        expires_in:
          type: integer
          description: TTL in seconds (900 = 15 minutes).
          example: 900

    # --- Policy --------------------------------------------------------------
    PolicyVersionInfo:
      type: object
      properties:
        id:
          type: string
        policy_group:
          type: string
        version:
          type: string
        checksum:
          type: string
        effective_from:
          type: string
          format: date-time
        effective_until:
          type: string
          format: date-time
          nullable: true
        is_active:
          type: boolean

    # --- Health --------------------------------------------------------------
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          example: rate_limit_exceeded
        message:
          type: string
          example: Token bucket exhausted. Retry after reset_at.
        details:
          type: object

# =============================================================================
# Paths
# =============================================================================
paths:
...
  /v1/decisions/{decision_id}:
    get:
      operationId: getDecision
      summary: Poll for an async decision result
      tags: [Tool Calls]
      parameters:
        - name: decision_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Decision is ready.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DecisionResponse"
        "202":
          description: Still pending.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PendingDecisionResponse"
        "404":
          description: Decision ID not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ---------------------------------------------------------------------------
  # Agent Registry
  # ---------------------------------------------------------------------------
  /v1/agents:
    post:
      operationId: registerAgent
      summary: Register a new agent and receive an API key
      description: |
        Protected by the root admin key (`SENTINEL_ADMIN_KEY`).
        The returned `api_key` is shown only once — store it securely.
      tags: [Agents]
      security:
        - AgentApiKey: []    # admin key required
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentRegistrationRequest"
      responses:
        "201":
          description: Agent registered. API key issued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentRegistrationResponse"
        "409":
          description: Agent with this name already exists in the tenant.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/agents/{agent_id}:
    get:
      operationId: getAgent
      summary: Get agent info and usage stats
      tags: [Agents]
      parameters:
        - name: agent_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentInfo"
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ---------------------------------------------------------------------------
  # Auth
  # ---------------------------------------------------------------------------
  /v1/auth/token:
    post:
      operationId: issueJWT
      summary: Exchange API key for a short-lived JWT
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TokenRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ---------------------------------------------------------------------------
  # Policies
  # ---------------------------------------------------------------------------
...
  /v1/policies/{policy_group}/activate:
    post:
      operationId: activatePolicy
      summary: Activate a specific policy version
      description: Deactivates the current version and activates the specified one.
      tags: [Policies]
      parameters:
        - name: policy_group
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [version]
              properties:
                version:
                  type: string
                  example: "3.1.0"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyVersionInfo"
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ---------------------------------------------------------------------------
  # WebSocket
  # ---------------------------------------------------------------------------
...